ThreadRename* moved to org.threadly.concurrent.wrapper.traceability
