+++ src/main/java/org/threadly/concurrent/wrapper/traceability/ThreadRenamingExecutorWrapper.java
@@ -1,20 +1,22 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper.traceability;
 
 import java.util.concurrent.Executor;
+
+import org.threadly.concurrent.AbstractSubmitterExecutor;
 
 /**
  * <p>Class which wraps a {@link Executor} and wraps all supplied tasks in a 
  * {@link ThreadRenamingRunnableWrapper}.  This allows you to make a pool where all tasks submitted 
  * inside it have the threads named in an identifiable way.</p>
  * 
- * @deprecated Moved to {@link org.threadly.concurrent.wrapper.traceability.ThreadRenamingExecutorWrapper}
- * 
  * @author jent
  * @since 4.3.0
  */
-@Deprecated
-public class ThreadRenamingExecutorWrapper 
-                 extends org.threadly.concurrent.wrapper.traceability.ThreadRenamingExecutorWrapper {
+public class ThreadRenamingExecutorWrapper extends AbstractSubmitterExecutor {
+  protected final Executor executor;
+  protected final String threadName;
+  protected final boolean replace;
+  
   /**
    * Constructs a new {@link ThreadRenamingExecutorWrapper}, wrapping a supplied {@link Executor}.  If 
    * {@code replace} is {@code false} the thread will be named such that 
@@ -25,6 +27,13 @@ public class ThreadRenamingExecutorWrapper
    * @param replace If {@code true} the original name wont be included in the thread name
    */
   public ThreadRenamingExecutorWrapper(Executor executor, String threadName, boolean replace) {
-    super(executor, threadName, replace);
+    this.executor = executor;
+    this.threadName = threadName;
+    this.replace = replace;
+  }
+
+  @Override
+  protected void doExecute(Runnable task) {
+    executor.execute(new ThreadRenamingRunnableWrapper(task, threadName, replace));
   }
 }
