+++ src/main/java/org/threadly/concurrent/wrapper/traceability/package-info.java
@@ -1,0 +1,6 @@
+
+/**
+ * <p>Wrappers for improving handling thread renaming during specific task (or wrapped pools) are 
+ * executed.</p>
+ */
+package org.threadly.concurrent.wrapper.traceability;
