+++ src/test/java/org/threadly/concurrent/wrapper/traceability/ThreadRenamingExecutorWrapperTest.java
@@ -1,8 +1,11 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper.traceability;
 
 import org.threadly.concurrent.PrioritySchedulerTest.PrioritySchedulerFactory;
+import org.threadly.concurrent.wrapper.traceability.ThreadRenamingExecutorWrapper;
+import org.threadly.concurrent.SubmitterExecutor;
+import org.threadly.concurrent.SubmitterExecutorInterfaceTest;
 
-@SuppressWarnings({"javadoc", "deprecation"})
+@SuppressWarnings("javadoc")
 public class ThreadRenamingExecutorWrapperTest extends SubmitterExecutorInterfaceTest {
   @Override
   protected SubmitterExecutorFactory getSubmitterExecutorFactory() {
