+++ src/test/java/org/threadly/concurrent/wrapper/traceability/ThreadRenamingSchedulerServiceWrapperTest.java
@@ -1,8 +1,13 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper.traceability;
 
 import org.threadly.concurrent.PrioritySchedulerTest.PrioritySchedulerFactory;
+import org.threadly.concurrent.wrapper.traceability.ThreadRenamingSchedulerServiceWrapper;
+import org.threadly.concurrent.SchedulerService;
+import org.threadly.concurrent.SchedulerServiceInterfaceTest;
+import org.threadly.concurrent.SubmitterExecutor;
+import org.threadly.concurrent.SubmitterScheduler;
 
-@SuppressWarnings({"javadoc", "deprecation"})
+@SuppressWarnings("javadoc")
 public class ThreadRenamingSchedulerServiceWrapperTest extends SchedulerServiceInterfaceTest {
   @Override
   protected SchedulerServiceFactory getSchedulerServiceFactory() {
