+++ src/test/java/org/threadly/concurrent/wrapper/traceability/ThreadRenamingSubmitterSchedulerWrapperTest.java
@@ -1,8 +1,12 @@
-package org.threadly.concurrent;
+package org.threadly.concurrent.wrapper.traceability;
 
 import org.threadly.concurrent.PrioritySchedulerTest.PrioritySchedulerFactory;
+import org.threadly.concurrent.wrapper.traceability.ThreadRenamingSubmitterSchedulerWrapper;
+import org.threadly.concurrent.SubmitterExecutor;
+import org.threadly.concurrent.SubmitterScheduler;
+import org.threadly.concurrent.SubmitterSchedulerInterfaceTest;
 
-@SuppressWarnings({"javadoc", "deprecation"})
+@SuppressWarnings("javadoc")
 public class ThreadRenamingSubmitterSchedulerWrapperTest extends SubmitterSchedulerInterfaceTest {
   @Override
   protected SubmitterSchedulerFactory getSubmitterSchedulerFactory() {
